REFACTOR : move redux type files
