+++ frontend/src/components/common/chatting/WaitingRoom.tsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from "react";
 import RoomCard from "components/common/chatting/RoomCard";
 import { ChatRoomDetail } from "types/Detail";
 import { useSelector, useDispatch } from "react-redux";
-import { CurrentChattingActionTypes } from "types/CurrentChatting";
+import { CurrentChattingActionTypes } from "types/redux/CurrentChatting";
 import CustomIconButton from "../utils/CustomIconButton";
 import "styles/Chatting.scss";
 import "styles/global.scss";
